+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Fixed
+- check-mysql-replication-status: fix code flow if server is not a slave (@DrMurx)
+
+### Changed
+- check-mysql-replication-status: refactoring & spec tests (@DrMurx)
+
+### Added
+- check-mysql-replication-status: added protection against `SHOW SLAVE STATUS` high lag reporting bug (@DrMurx)
 
 ## [3.1.1] - 2019-03-04
 ### Fixed